data type change

+++  cf-A - Candies.cpp
@@ -1,24 +1,24 @@
 #include <bits/stdc++.h>
 using namespace std;
-#define ll long long int
+#define int long long int
 #define pb push_back
 #define fastIO ios_base::sync_with_stdio(false);cin.tie(NULL);
 #define endl "\n"
 
-int main() 
+int32_t main() 
 {
     #ifndef ONLINE_JUDGE
         freopen("input.txt", "r", stdin);
         freopen("output.txt", "w", stdout);
     #endif
     fastIO;
-    ll t,n,a,i;
+    int t,n,a,i;
     cin>>t;
     while(t--)
     {
         cin>>n;
-        ll mul=2;
-        ll add=3;
+        int mul=2;
+        int add=3;
         while(n%add!=0)
         {
             mul*=2;
